+++ R/zzz.R
@@ -4,8 +4,7 @@ CONFIG_FILENAME <- "config.yaml" # staged.dependencies._storage_dir
 # Use this function when you want to change the storage directory while the package is already loaded
 # It does not delete the old storage directory
 setup_storage_dir <- function(storage_dir) {
-  options(staged.dependencies._storage_dir = storage_dir)
-
+  set_storage_dir(storage_dir)
   # only copy config if storage dir does not exist
   if (!dir.exists(storage_dir)) {
     dir.create(storage_dir)
@@ -23,7 +22,16 @@ get_storage_dir <- function() {
   options()$staged.dependencies._storage_dir
 }
 
+set_storage_dir <- function(storage_dir) {
+  options("staged.dependencies._storage_dir" = storage_dir)
+}
+
 .onLoad <- function(libname, pkgname) {
+
+  storage_dir <- getOption(
+    "staged.dependencies._storage_dir",
+    path.expand("~/.staged.dependencies")
+  )
   op <- options()
   op.package <- list(
     # mapping from hosts to
@@ -34,5 +42,5 @@ get_storage_dir <- function() {
   )
   toset <- !(names(op.package) %in% names(op))
   if (any(toset)) options(op.package[toset])
-  setup_storage_dir(path.expand("~/.staged.dependencies"))
+  setup_storage_dir(storage_dir)
 }