+++ tests/testthat/test-zzz.R
@@ -10,7 +10,8 @@ test_that("setup_storage_dir works", {
 
   # modify storage directory to a new random one
   new_storage_dir <- tempfile("random_storage_dir")
-  on.exit(setup_storage_dir(TESTS_STORAGE_DIR), add = TRUE)
+  set_storage_dir(new_storage_dir)
+  on.exit({setup_storage_dir(TESTS_STORAGE_DIR)}, add = TRUE)
   setup_storage_dir(new_storage_dir)
 
   expect_equal(get_storage_dir(), new_storage_dir)
@@ -22,4 +23,15 @@ test_that("setup_storage_dir works", {
     get_packages_cache_dir(),
     file.path(new_storage_dir, "packages_cache")
   )
+})
+
+test_that("get/set_storage_dir", {
+  dir <- tempdir()
+
+  withr::with_options(list("staged.dependencies._storage_dir" = dir), {
+    expect_equal(get_storage_dir(), dir)
+  })
+
+  set_storage_dir(dir)
+  expect_equal(get_storage_dir(), dir)
 })